+++ Publish_Weaver/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.0
 * Optimized fractionators.
+* Additional assembler data access optimizations.
 
 ## 1.0.1
 * Fixed sorters on active planets would be updated twice per tick which would sometimes result in a `DivideByZeroException`.